Major Improvements and Test Additions
Modified the random generator so the seed could be changed.
Changed the way in which the grid gets updated. Now, updates to the grid are initiated by the Cell object.
Overhauled the way adjacent cells are retrieved. Now, all of the 8 adjacent cells are returned from a function as a list.
Added basic user input and game flow.

Added tests for random seed generation, to make sure that when a preset seed is used, a cells properties are set correctly.